chore: move internal/connection
