+++ internal/connection/state.go
