+++ internal/connection/state_test.go
