+++ messaging/waku/nwaku.go
@@ -59,8 +59,8 @@ import (
 	wakutimesource "github.com/waku-org/waku-go-bindings/waku/timesource"
 
 	gocommon "github.com/status-im/status-go/common"
-	"github.com/status-im/status-go/connection"
 	cryptotypes "github.com/status-im/status-go/crypto/types"
+	"github.com/status-im/status-go/internal/connection"
 	"github.com/status-im/status-go/internal/timesource"
 	"github.com/status-im/status-go/logutils"
 	"github.com/status-im/status-go/messaging/waku/common"
