+++ messaging/waku/types/waku.go
@@ -17,7 +17,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/p2p/enode"
 
-	"github.com/status-im/status-go/connection"
+	"github.com/status-im/status-go/internal/connection"
 )
 
 type ConnStatus struct {
